requirements.txt
albumentations==1.3.1
lmdb==1.3.0
numpy==1.21.6
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.5.4.60
Pillow==10.1.0
scikit_learn==0.23.1
scikit_learn==1.0.2
scipy==1.7.3
tensorboardX==2.6.2.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.65.0
